Global: expands pid to i64
See `process.rs` for detail

+++ src/columns/file_name.rs
@@ -14,8 +14,8 @@ use std::path::Path;
 pub struct FileName {
     header: String,
     unit: String,
-    fmt_contents: HashMap<i32, String>,
-    raw_contents: HashMap<i32, String>,
+    fmt_contents: HashMap<i64, String>,
+    raw_contents: HashMap<i64, String>,
     width: usize,
 }
