+++ src/columns/user_real.rs
@@ -8,8 +8,8 @@ use uzers::Users;
 pub struct UserReal {
     header: String,
     unit: String,
-    fmt_contents: HashMap<i32, String>,
-    raw_contents: HashMap<i32, String>,
+    fmt_contents: HashMap<i64, String>,
+    raw_contents: HashMap<i64, String>,
     width: usize,
 }
